Tema
Proiect

+++ src/hotelManagementSystem/Apartment.java
@@ -4,7 +4,7 @@ public class Apartment extends Room {
     private String dotateKitchen;
     private int numberRooms;
 
-    public Apartment(int number, String typeOfRoom, String view, String occupiedStatus, int price, String dotateKitchen, int numberRooms) {
+    public Apartment(int number, String typeOfRoom, String view, String occupiedStatus, double price, String dotateKitchen, int numberRooms) {
         super(number, typeOfRoom, view, occupiedStatus, price);
         this.dotateKitchen = dotateKitchen;
         this.numberRooms = numberRooms;
